#2639 Reapplied GPU benchmark fix
First test returns quarter to a half the throughput, do two tests.
May be caused by driver, may be some 'energy saving', but not important
enough to spend time investingating. It was working the same way prior
to ExtraFPS, but viewer was running an extra CPU test that 'preheated'
the system.

Also increasing minimum throughput as numerous new features, like
mirrors and pbr were added and requirements are now higher.